Create Spotify Analysis
I conducted an analysis of Spotify's 2021 Top 50 data, which I downloaded from Kaggle. My process began by setting up a table in SQLiteStudio and importing the data. Once the data was ready, I used a series of SQL queries to explore various aspects of the songs and artists.

My analysis focused on several key areas. I looked at popularity trends to find the top and least popular songs, as well as the artists with the most songs in the top half of the chart. I also delved into the audio features of the songs, using metrics like danceability, energy, and loudness to understand the characteristics of the music. Additionally, I investigated song length and used more advanced queries to explore correlations, like whether songs with a higher speechiness score also tended to be more popular. My final product is a collection of well-structured SQL queries that answer these questions, showcasing my ability to use SQL for data analysis and uncover meaningful insights from a dataset.